datatypes: Allow zeroing out buffer contents before finishing
With the SOL_BUFFER_CLEAR_MEMORY flag, the buffer will be cleared out
whenever finishing it.

If this flag is specified, SOL_BUFFER_FLAGS_MEMORY_NOT_OWNED is
implicitly specified.  The reason is that, since `realloc()` is used
internally in the sol_buffer, and if the block is shrinking (so old
contents will still be in memory), or when the pointer changes (and
realloc() frees it internally), old contents will still linger in
memory.  Would have to detect these cases, which is kind of hard to do
when realloc() returns a new pointer, since the old pointer is unusable
when it returns and there's no way to know if realloc() will return a
new pointer.  The easiest way out is to ensure that realloc() is never
used by requiring these flags to be set.  It's possible to create a
realloc()-like function for this purpose, but for simplicity sake,
let's keep it like this.

Signed-off-by: Leandro Pereira <leandro.pereira@intel.com>